feat: 피그마 적용
