--- src/assets/component/button-3d.css
@@ -1,84 +1,0 @@
-@media screen and (min-width: 768px) {
-  .content-3d {
-    margin: auto;
-    text-align: center;
-  }
-
-  a[class*="btn"] {
-    text-decoration: none;
-  }
-  .btn-3d {
-    position: relative;
-    display: inline-block;
-    font-size: 2vw;
-    padding: 1.2% 3.2%;
-    color: white;
-    border-radius: 6px;
-    border: 0.7px solid black;
-    text-align: center;
-    transition: top 0.01s linear;
-    text-shadow: 0 1px 0 rgba(0, 0, 0, 0.15);
-  }
-  .btn-3d.yellow:hover {
-    background-color: #f0d264;
-  }
-
-  .btn-3d:active {
-    top: 9px;
-  }
-
-  .btn-3d.yellow {
-    background-color: #f0d264;
-    box-shadow: 0 0 0 1px #f0d264 inset,
-      0 0 0 2px rgba(255, 255, 255, 0.15) inset,
-      0 8px 0 0 rgba(196, 172, 83, 0.7), 0 8px 0 1px rgba(0, 0, 0, 0.4),
-      0 8px 8px 1px rgba(0, 0, 0, 0.5);
-  }
-  .btn-3d.yellow:active {
-    box-shadow: 0 0 0 1px #f0d264 inset,
-      0 0 0 2px rgba(255, 255, 255, 0.15) inset, 0 0 0 1px rgba(0, 0, 0, 0.4);
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .content-3d {
-    margin: auto;
-    text-align: center;
-  }
-
-  a[class*="btn"] {
-    text-decoration: none;
-  }
-
-  .btn-3d {
-    width: 50%;
-    position: relative;
-    font-size: 5.5vw;
-    padding: 20px 60px;
-    color: white;
-    border-radius: 6px;
-    border: 0.7px solid black;
-    text-align: center;
-    transition: top 0.01s linear;
-    text-shadow: 0 1px 0 rgba(0, 0, 0, 0.15);
-  }
-  .btn-3d.yellow:hover {
-    background-color: #f0d264;
-  }
-
-  .btn-3d:active {
-    top: 9px;
-  }
-
-  .btn-3d.yellow {
-    background-color: #f0d264;
-    box-shadow: 0 0 0 1px #f0d264 inset,
-      0 0 0 2px rgba(255, 255, 255, 0.15) inset,
-      0 8px 0 0 rgba(196, 172, 83, 0.7), 0 8px 0 1px rgba(0, 0, 0, 0.4),
-      0 8px 8px 1px rgba(0, 0, 0, 0.5);
-  }
-  .btn-3d.yellow:active {
-    box-shadow: 0 0 0 1px #f0d264 inset,
-      0 0 0 2px rgba(255, 255, 255, 0.15) inset, 0 0 0 1px rgba(0, 0, 0, 0.4);
-  }
-}
