--- src/assets/component/button.css
@@ -1,89 +1,0 @@
-.button {
-  all: unset;
-  background-color: #ac59ff;
-  border-radius: 30px;
-  box-sizing: border-box;
-  height: 47px;
-  position: relative;
-  width: 280px;
-}
-
-.button .text-wrapper {
-  color: #ffffff;
-  font-family: "Pretendard-SemiBold", Helvetica;
-  font-size: 16px;
-  font-weight: 600;
-  left: 112px;
-  letter-spacing: 0;
-  line-height: normal;
-  position: absolute;
-  text-align: center;
-  top: 14px;
-  white-space: nowrap;
-  width: 56px;
-}
-
-/* @media screen and (min-width: 768px) {
-  .btnBox {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .btn {
-    padding: 1.2% 3.2%;
-    font-size: 2vw;
-    text-transform: uppercase;
-    letter-spacing: 2.5px;
-    font-weight: 500;
-    color: #000;
-    background-color: #fff;
-    border: none;
-    border-radius: 45px;
-    box-shadow: 0px 8px 15px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s ease 0s;
-    cursor: pointer;
-    outline: none;
-  }
-
-  .btn:hover {
-    background-color: #2ee59d;
-    box-shadow: 0px 15px 20px rgba(46, 229, 157, 0.4);
-    color: #fff;
-    transform: translateY(-7px);
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .btnBox {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .btn {
-    padding: 4% 6%;
-    font-size: 8vw;
-    text-transform: uppercase;
-    text-decoration: none;
-    letter-spacing: 2.5px;
-    font-weight: 50%;
-    color: #000;
-    background-color: #fff;
-    border: none;
-    border-radius: 45px;
-    box-shadow: 0px 8px 15px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s ease 0s;
-    cursor: pointer;
-    outline: none;
-  }
-
-  .btn:hover {
-    background-color: #2ee59d;
-    box-shadow: 0px 15px 20px rgba(46, 229, 157, 0.4);
-    color: #fff;
-    transform: translateY(-7px);
-  }
-} */
