--- src/assets/pages/mainpage.css
@@ -1,43 +1,0 @@
-@media screen and (min-width: 768px) {
-
-  body {
-    margin: 0;
-  }
-  .mainContainer {
-    position: fixed;
-    width: 100%;
-    height: 100vh;
-    justify-content: center;
-    align-items: center;
-    overflow: hidden;
-    background: #CBC0D3 no-repeat center;
-    background-image: url("../../../public/background_mobile.gif");
-    background-size: cover;
-  }
-
-  .main__btnBox {
-    margin: 12% auto;
-  }
-
-  /* .main__btn {
-    border: 5px solid black;
-  } */
-  /* .main__btnBox {
-    margin: 34% auto;
-  } */
-}
-
-@media screen and (max-width: 768px) {
-  .mainContainer {
-    width: 100%;
-    height: 100vh;
-    position: fixed;
-    overflow: hidden;
-    background: #CBC0D3 no-repeat center;
-    background-image: url("../../../public/background_mobile.gif");
-    background-size: cover;
-  }
-  .main__btnBox {
-    margin: 25% auto;
-  }
-}
