--- src/assets/pages/normalImageResult.css
@@ -1,34 +1,0 @@
-@media screen and (min-width: 768px) {
-    
-.result__box {
-    /* display: flex; */
-    width: 80%;
-    /* height: 80vh; */
-        margin: 3% auto;
-        border-radius: 20px;
-        border: 10px solid #2ee59d;
-        overflow: auto;
-    }
-
-.result_image {
-    object-fit: cover;
-}
-}
-
-@media screen and (max-width: 768px) {
-
-    .result__box {
-        overflow: auto;
-        /* width: 100%; */
-        /* height: 50vh; */
-        margin: 10% auto;
-        
-        border-radius: 20px;
-        border: 10px solid #2ee59d
-    }
-    
-    .result_image {
-        object-fit: cover;
-    }
-    
-}
