+++ src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import "./assets/main.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
