+++ src/layouts/Layouts.js
@@ -1,13 +1,13 @@
 import { useState } from "react";
 import { Outlet } from "react-router-dom";
-import Header from "../components/header/Header";
+// import Header from "../component/header/Header";
 // import LoginModal from "../component/login/L doginModal";
 
 function Layout() {
   const [edit, setEdit] = useState(false);
   return (
     <div>
-      <Header edit={edit} setEdit={setEdit} />
+      {/* <Header edit={edit} setEdit={setEdit} /> */}
       {/* <LoginModal edit={edit} setEdit={setEdit} /> */}
       <Outlet />
     </div>
