+++ src/pages/mainpage/mainPage.css
@@ -1,0 +1,140 @@
+.index {
+  background-color: #ffffff;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  width: 100%;
+}
+
+.index .overlap-group-wrapper {
+  background-color: #ffffff;
+  height: 652px;
+  width: 400px;
+}
+
+.index .overlap-group {
+  height: 618px;
+  position: relative;
+}
+
+.index .image {
+  height: 599px;
+  left: 0;
+  object-fit: cover;
+  position: absolute;
+  top: 0;
+  width: 400px;
+}
+
+.index .rectangle {
+  background: linear-gradient(
+    180deg,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 0.8) 43.54%,
+    rgb(255, 255, 255) 69.06%
+  );
+  height: 166px;
+  left: 0;
+  position: absolute;
+  top: 439px;
+  width: 400px;
+}
+
+.index .button-1 {
+  left: 60px !important;
+  position: absolute !important;
+  top: 508px !important;
+}
+
+.index .button-instance {
+  left: 67px !important;
+  width: unset !important;
+}
+
+.index .text-top {
+  height: 61px;
+  left: 94px;
+  position: absolute;
+  top: 51px;
+  width: 217px;
+}
+
+.index .text-wrapper-2 {
+  color: #ffffff;
+  font-family: "Pretendard-Regular", Helvetica;
+  font-size: 24px;
+  font-weight: 400;
+  left: 55px;
+  letter-spacing: 0;
+  line-height: normal;
+  position: absolute;
+  top: 0;
+}
+
+.index .AI {
+  color: #ffffff;
+  font-family: "Pretendard-Regular", Helvetica;
+  font-size: 12px;
+  font-weight: 400;
+  left: 0;
+  letter-spacing: 0;
+  line-height: normal;
+  position: absolute;
+  text-align: center;
+  top: 33px;
+}
+
+.index .button-1-instance {
+  left: 60px !important;
+  position: absolute !important;
+  top: 571px !important;
+}
+
+.index .design-component-instance-node {
+  left: 95px !important;
+  width: unset !important;
+}
+
+.index .view {
+  left: 164px !important;
+  position: absolute !important;
+  top: 459px !important;
+}
+
+/* @media screen and (min-width: 768px) {
+
+  body {
+    margin: 0;
+  }
+  .mainContainer {
+    position: fixed;
+    width: 100%;
+    height: 100vh;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+    background: #CBC0D3 no-repeat center;
+    background-image: url("../../../public/background_mobile.gif");
+    background-size: cover;
+  }
+
+  .main__btnBox {
+    margin: 12% auto;
+  }
+
+}
+
+@media screen and (max-width: 768px) {
+  .mainContainer {
+    width: 100%;
+    height: 100vh;
+    position: fixed;
+    overflow: hidden;
+    background: #CBC0D3 no-repeat center;
+    background-image: url("../../../public/background_mobile.gif");
+    background-size: cover;
+  }
+  .main__btnBox {
+    margin: 25% auto;
+  }
+} */
